third_party/blink/web_tests/wpt_internal/attribution-reporting/aggregatable-report-no-contributions.sub.https.html

<!doctype html>
<meta charset=utf-8>
<meta name=timeout content=long>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/helpers.js"></script>
<script>

attribution_reporting_promise_test(async t => {
  const host = `https://{{host}}`;

  await registerAttributionSrc(t, {
    source: {
      destination: host,
      source_event_id: '123',
      aggregation_keys: {
        campaignCounts: '0x159',
      },
      debug_key: '456',
    },
    cookie: attributionDebugCookie,
  });

  await registerAttributionSrc(t, {
    trigger : {
      // Create event-level report to avoid verbose debug report created for
      // event-level attribution.
      event_trigger_data: [{}],
      aggregatable_values: {
        geoValue: 32768,
      },
      debug_key: '654',
      debug_reporting: true,
    },
    cookie: attributionDebugCookie,
  });

  const debugPayload = await pollVerboseDebugReports();
  assert_equals(debugPayload.reports.length, 1);
  const debugReport = JSON.parse(debugPayload.reports[0].body);
  assert_equals(debugReport.length, 1);
  assert_equals(debugReport[0].type, 'trigger-aggregate-no-contributions');
  assert_own_property(debugReport[0], 'body');
  const debugReportBody = debugReport[0].body;
  assert_equals(debugReportBody.attribution_destination, host);
  assert_equals(debugReportBody.source_event_id, '123');
  assert_equals(debugReportBody.source_site, host);
  assert_equals(debugReportBody.source_debug_key, '456');
  assert_equals(debugReportBody.trigger_debug_key, '654');
}, 'Aggregatable report is not created due to no contributions.');
</script>
